Marginal UTF8 support
The text renderer heavily failed on UTF8 chars before, now it renders at
least some of them ok :)

* TTF module exposes u16 codepoints, do a bunch of u8 -> u16 to support
  the larger ranges
* Parse utf8 codepoints as well as we can using zig's builtin unicode
  support
* When we see codepoints that are too large, fall back to the
  replacement character

The u16 restriction is a limitation of our ttf implementation. There are
cmap subtables which define larger ranges, we just don't look at them
yet

We have made no attempt to do anything other than use each utf8
codepoint directly. We surely are not handling compound glyphs or
anything correctly. This is still a lot better than nothing though :)

+++ src/sphtext/src/GlyphAtlas.zig
@@ -20,7 +20,7 @@ texture: Texture,
 tex_width: u31,
 tex_height: u31,
 //FIXME: Utf8 codepoints
-glyph_locations: std.AutoHashMap(u8, UVBBox),
+glyph_locations: std.AutoHashMap(u16, UVBBox),
 x_cursor_px: u16 = 0,
 y_cursor_px: u16 = 0,
 row_max_height: u16 = 0,
@@ -75,7 +75,7 @@ pub fn init(gpa: Allocator, gl_alloc: *GlAlloc) !GlyphAtlas {
 
     return .{
         .program = program,
-        .glyph_locations = std.AutoHashMap(u8, UVBBox).init(gpa),
+        .glyph_locations = std.AutoHashMap(u16, UVBBox).init(gpa),
         .glyph_render_source = render_source,
         .texture = texture,
         .tex_width = tex_width,
@@ -92,7 +92,7 @@ pub fn getGlyphLocation(
     self: *GlyphAtlas,
     scratch_alloc: *ScratchAlloc,
     scratch_gl: *GlAlloc,
-    char: u8,
+    char: u16,
     point_size: f32,
     ttf: ttf_mod.Ttf,
     distance_field_renderer: sphrender.DistanceFieldGenerator,
@@ -111,7 +111,7 @@ pub fn getGlyphLocation(
     return gop.value_ptr.*;
 }
 
-fn addCharToAtlas(self: *GlyphAtlas, scratch_alloc: *ScratchAlloc, scratch_gl: *GlAlloc, char: u8, point_size: f32, ttf: ttf_mod.Ttf, distance_field_renderer: sphrender.DistanceFieldGenerator) !UVBBox {
+fn addCharToAtlas(self: *GlyphAtlas, scratch_alloc: *ScratchAlloc, scratch_gl: *GlAlloc, char: u16, point_size: f32, ttf: ttf_mod.Ttf, distance_field_renderer: sphrender.DistanceFieldGenerator) !UVBBox {
     const gl_checkpoint = scratch_gl.checkpoint();
     defer scratch_gl.restore(gl_checkpoint);
 
@@ -294,7 +294,7 @@ fn makeDistanceField(
     scratch_alloc: *ScratchAlloc,
     gl_alloc: *GlAlloc,
     ttf: ttf_mod.Ttf,
-    c: u8,
+    c: u16,
     width: usize,
     height: usize,
     mask_texture: sphrender.Texture,
